chore: refactor controllers and services
