+++ backend/src/controllers/app.ts
@@ -1,56 +1,18 @@
-import { Body, Controller, Get, Param, Post, Query } from '@nestjs/common';
-import { parseEther } from 'viem';
+import { Controller, Get, Query } from '@nestjs/common';
 
 import { AppService } from '../services/app';
-import { MintTokenDto } from '../dtos/mintToken.dto';
 
 @Controller()
 export class AppController {
   constructor(private readonly appService: AppService) {}
-
-  @Get('contract-address')
-  getContractAddress() {
-    return { result: this.appService.getContractAddress() };
-  }
-
-  @Get('token-name')
-  async getTokenName() {
-    return { result: await this.appService.getTokenName() };
-  }
-
-  @Get('total-supply')
-  async getTotalSupply() {
-    return { result: await this.appService.getTotalSupply() };
-  }
-
-  @Get('token-balance/:address')
-  async getTokenBalance(@Param('address') address: string) {
-    return { result: await this.appService.getTokenBalance(address) };
-  }
 
   @Get('transaction-receipt')
   async getTransactionReceipt(@Query('hash') hash: string) {
     return { result: await this.appService.getTransactionReceipt(hash) };
   }
 
-  // MINT TOKENS
   @Get('server-wallet-address')
   getServerWalletAddress() {
     return { result: this.appService.getServerWalletAddress() };
-  }
-
-  @Get('check-minter-role')
-  checkMinterRole() {
-    return { result: this.appService.checkMinterRole() };
-  }
-
-  @Post('mint-tokens')
-  async mintTokens(@Body() body: MintTokenDto) {
-    return {
-      result: this.appService.mintTokens(
-        body.address,
-        parseEther(`${body.amountToMint}`),
-      ),
-    };
   }
 }
