+++ backend/src/main.ts
@@ -16,7 +16,8 @@ async function bootstrap() {
     .setTitle('TokenizedBallot API')
     .setDescription('The API description')
     .setVersion('1.0')
-    .addTag('tokenizedBallot')
+    .addTag('MyToken')
+    .addTag('TokenizedBallot')
     .build();
   const document = SwaggerModule.createDocument(app, config);
   SwaggerModule.setup('api', app, document);
